split combined queries into their own AST since they have weird additional requirements
